+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.0.42 (2026-01-17)
+
+- **Summary:** PERSONALITIES (PRISM): Update Source
+- **Files:**
+  - `Source/PluginProcessor.cpp`
+  - `Source/PluginProcessor.h`
+
 ## v0.0.41 (2026-01-17)
 
 - **Summary:** PERSONALITIES (PRISM): Update Source